requirements.txt
sous-chef @ git+https://github.com/mediacloud/sous-chef@v0.2
statsd-client==1.0.*